+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Upcoming v0.6.0: Revert to basic ruby Logger
   - Now, rails app can assign `TusClient.logger = Rails.logger`
   - still logs :source
 - Logger.level = Logger::INFO (was ERROR)
+- add binstub `bin/rackup`
 
 2020-08-06 v0.5.0: Logged 'source' is tus code. Namespace support classes.
 - Move Responsorial to TusClient namespace